backend/requirements.txt

fastapi==0.109.2
uvicorn==0.27.1
apify-client==1.6.0
httpx==0.27.0
pandas==2.2.0
python-dotenv==1.0.1
